trying to implement user id

+++ Capabilities/chalicelib/dynamo_service.py
@@ -4,6 +4,8 @@ class DynamoService:
     def __init__(self):
         dynamodb = boto3.resource('dynamodb')
         self.table = dynamodb.Table('contacts')
+        self.dynamodb_client = boto3.client('dynamodb')
+        self.user_id = ""
 
     def save_item(self,data):
 
@@ -13,9 +15,13 @@ class DynamoService:
     
     def delete_contact(self, lead_name):
 
-            self.table.delete_item(Key={'lead_name': lead_name})
+            self.table.delete_item(Key={'user_id':user_id,'lead_name': lead_name})
         
     def update_contact(self, data):
+
+        user_id = data['user_id']
+        #if user_id == to user_id in da
+        #if user_id 
 
         print("In update contact")
         response = self.table.update_item(
@@ -28,7 +34,7 @@ class DynamoService:
     def query_contacts(self, lead_name):
 
         response = self.table.query(KeyConditionExpression=Key("lead_name").eq(lead_name))
-        
+
         items = response['Items']
         print(items)
         return(items)
